Added Code to install libmodbus through yum package manager only if OS is RedHat/CentOS 7

+++ requirements.sh
@@ -7,11 +7,17 @@ else
 	which yum >/dev/null 2>&1
 	if [ $? -eq 0 ]; then
 		sudo yum -y install epel-release
-		# Package of libmodbus is not available for CentOS Stream 9
-		# sudo yum install -y libmodbus
-		# sudo yum install -y libmodbus-devel
-		cd ${HOME} && wget https://libmodbus.org/releases/libmodbus-3.0.8.tar.gz && tar -xvzf libmodbus-3.0.8.tar.gz && \
-		cd libmodbus-3.0.8/ && ./configure && make && sudo make install && \
-		cd ${HOME} && sudo rm -rf libmodbus-3.0.8*
+		# Check OS NAME and VERSION
+		OS_NAME=$(grep -o '^NAME=.*' /etc/os-release | cut -f2 -d\" | sed 's/"//g')
+		OS_VERSION=$(grep -o '^VERSION_ID=.*' /etc/os-release | cut -f2 -d\" | sed 's/"//g')
+		# Install libmodbus thorugh yum package only if 
+		if [[ ( ${OS_NAME} == *"Red Hat"* || ${OS_NAME} == *"CentOS"* ) && ${OS_VERSION} == *"7"* ]]; then
+			sudo yum install -y libmodbus
+			sudo yum install -y libmodbus-devel
+		elif [[ ( ${OS_NAME} == *"Red Hat"* || ${OS_NAME} == *"CentOS"* ) && ${OS_VERSION} >= *"9"* ]]; then
+			cd ${HOME} && wget https://libmodbus.org/releases/libmodbus-3.0.8.tar.gz && tar -xvzf libmodbus-3.0.8.tar.gz && \
+			cd libmodbus-3.0.8/ && ./configure && make && sudo make install && \
+			cd ${HOME} && sudo rm -rf libmodbus-3.0.8*
+		fi
 	fi
 fi
